+++ templates/mca/multi/MCA.Api/Program.cs
@@ -3,15 +3,11 @@ using Serilog;
 #endif
 using Scalar.AspNetCore;
 using MCA.Application;
-using MCA.Application.Services;
 using MCA.Domain.Interfaces;
 using MCA.Infrastructure.Data;
 using MCA.Infrastructure.Repositories;
-using MCA.Infrastructure.Services;
 using MCA.Application.Commands;
-#if (UseAuth)
 using MCA.Application.Handlers;
-#endif
 using MCA.Api.Endpoints;
 using Microsoft.EntityFrameworkCore;
 using MinimalCleanArch.DataAccess.Repositories;
@@ -50,6 +46,7 @@ using MCA.Api.Configuration;
 using MCA.Api.Services;
 using MCA.Application.Interfaces;
 using MCA.Infrastructure.Configuration;
+using MCA.Infrastructure.Services;
 #endif
 #if (UseMessaging)
 using Wolverine;
@@ -154,8 +151,8 @@ builder.Services.AddDbContext<AppDbContext>((sp, options) =>
 builder.Services.AddScoped<ITodoRepository>(sp => new TodoRepository(sp.GetRequiredService<AppDbContext>()));
 builder.Services.AddScoped<IUnitOfWork>(sp => new UnitOfWork(sp.GetRequiredService<AppDbContext>()));
 
-// Services
-builder.Services.AddScoped<ITodoService, TodoService>();
+// Application handlers (use cases)
+builder.Services.AddScoped<TodoCommandHandler>();
 #if (UseAudit)
 builder.Services.AddHttpContextAccessor();
 #endif