+++ templates/mca/multi/MCA.Application/Handlers/ResetPasswordHandler.cs
@@ -1,5 +1,5 @@
 using MCA.Application.Commands;
-using MCA.Domain.Entities;
+using MCA.Application.Identity;
 using Microsoft.AspNetCore.Identity;
 using MinimalCleanArch.Domain.Common;